+++ modules/client/examples/01-client-token/00-installation.ts
@@ -11,13 +11,15 @@ underlying network requests.
 */
 import {
   Client,
-  TokenVotingClient,
   Context,
   DaoCreationSteps,
   GasFeeEstimation,
   ICreateParams,
   ITokenVotingPluginInstall,
+  TokenVotingClient,
+  VotingMode,
 } from "@aragon/sdk-client";
+import { IMetadata } from "../../src";
 import { contextParams } from "../00-client/00-context";
 
 const context: Context = new Context(contextParams);
@@ -27,10 +29,12 @@ const client: Client = new Client(context);
 
 // Use an already existing ERC20 token
 const pluginInitParams1: ITokenVotingPluginInstall = {
-  settings: {
+  votingSettings: {
     minDuration: 60 * 60 * 24 * 2, // seconds
-    minTurnout: 0.25, // 25%
-    minSupport: 0.5, // 50%
+    minParticipation: 0.25, // 25%
+    supportThreshold: 0.5, // 50%
+    minProposerVotingPower: BigInt("5000"), // default 0
+    votingMode: VotingMode.STANDARD, // default standard
   },
   useToken: {
     address: "0x...",
@@ -38,10 +42,12 @@ const pluginInitParams1: ITokenVotingPluginInstall = {
 };
 // Mint a new token
 const pluginInitParams2: ITokenVotingPluginInstall = {
-  settings: {
-    minDuration: 60 * 60 * 24, // seconds
-    minTurnout: 0.4, // 40%
-    minSupport: 0.55, // 55%
+  votingSettings: {
+    minDuration: 60 * 60 * 24 * 2, // seconds
+    minParticipation: 0.25, // 25%
+    supportThreshold: 0.5, // 50%
+    minProposerVotingPower: BigInt("5000"), // default 0
+    votingMode: VotingMode.EARLY_EXECUTION, // default standard
   },
   newToken: {
     name: "Token",
@@ -64,15 +70,29 @@ const pluginInitParams2: ITokenVotingPluginInstall = {
     ],
   },
 };
-const tokenVotingInstallPluginItem1 = TokenVotingClient.encoding.getPluginInstallItem(
-  pluginInitParams1,
-);
-const tokenVotingInstallPluginItem2 = TokenVotingClient.encoding.getPluginInstallItem(
-  pluginInitParams2,
-);
+const tokenVotingInstallPluginItem1 = TokenVotingClient.encoding
+  .getPluginInstallItem(
+    pluginInitParams1,
+  );
+const tokenVotingInstallPluginItem2 = TokenVotingClient.encoding
+  .getPluginInstallItem(
+    pluginInitParams2,
+  );
+
+const daoMetadata: IMetadata = {
+  name: "My DAO",
+  description: "This is a description",
+  avatar: "",
+  links: [{
+    name: "Web site",
+    url: "https://...",
+  }],
+};
+
+const metadataUri = await client.methods.pinMetadata(daoMetadata);
 
 const createParams: ICreateParams = {
-  metadataUri: "ipfs://123456789",
+  metadataUri,
   ensSubdomain: "my-org", // my-org.dao.eth
   plugins: [tokenVotingInstallPluginItem1, tokenVotingInstallPluginItem2],
 };