+++ modules/client/examples/03-encoders/07-update-plugin-settings-token.ts
@@ -2,10 +2,11 @@
 ### Set Plugin Config (TokenVoting)
 */
 import {
-  TokenVotingClient,
   Context,
   ContextPlugin,
-  IPluginSettings,
+  TokenVotingClient,
+  VotingMode,
+  VotingSettings,
 } from "@aragon/sdk-client";
 import { contextParams } from "../00-client/00-context";
 
@@ -14,10 +15,12 @@ const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
 const client = new TokenVotingClient(contextPlugin);
 
 // create config action
-const configActionPrarms: IPluginSettings = {
-  minDuration: 60 * 60 * 24,
-  minSupport: 0.3, // 30%
-  minTurnout: 0.5, // 50%
+const configActionPrarms: VotingSettings = {
+  minDuration: 60 * 60 * 24 * 2, // seconds
+  minParticipation: 0.25, // 25%
+  supportThreshold: 0.5, // 50%
+  minProposerVotingPower: BigInt("5000"), // default 0
+  votingMode: VotingMode.STANDARD, // default standard
 };
 
 const pluginAddress = "0x1234567890123456789012345678901234567890";