+++ modules/client/CHANGELOG.md
@@ -49,6 +49,13 @@ TEMPLATE:
 - renames `QueryToken` to `QueryTokenVotingToken`
 - renames `toErc20Proposal` to `toTokenVotingProposal`
 - renames `toErc20ProposalListItem` to `toTokenVotingProposalListItem`
+- renames `IPluginSettings` to `VotingSettings`
+- renames `minTurnout` to `minParticipation`
+- renames `minSupport` to `supportThreshold`
+- add `minProposerVotingPower`, `earlyExecution`and `voteReplacement` fields to `VotingSettings` and graphql responses
+- update `ethers-contracts` version to `0.5.0-alpha`
+- rename `getSettings` to `getVotingSettings`
+- rename `encodeUpdatePluginSettingsAction` to `encodeUpdateVotingSettingsAction`
 
 
 ## [0.16.2-alpha]